+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - [5937](https://github.com/vegaprotocol/vega/issues/5937) - Add more flexibility to market creation bonus 
 - [5708](https://github.com/vegaprotocol/vega/issues/5708) - Use market price when reporting average trade price
 - [5949](https://github.com/vegaprotocol/vega/issues/5949) - Transfers processed in the order they were received
+- [5966](https://github.com/vegaprotocol/vega/issues/5966) - Do not send transaction from wallet if `chainID` is empty
 - [5675](https://github.com/vegaprotocol/vega/issues/5675) - Add transaction information to all database tables
 - [6004](https://github.com/vegaprotocol/vega/issues/6004) - Probability of trading refactoring
 - [5849](https://github.com/vegaprotocol/vega/issues/5849) - Use network parameter from creation time of the proposal for requirements